test: add tests for node 18, node 19 and vite-node
