+++ tests/package.json
@@ -1,0 +1,7 @@
+{
+  "private": true,
+  "type": "module",
+  "devDependencies": {
+    "iron-webcrypto": "workspace:*"
+  }
+}
